Refactor file structure by moving recentfiles and settings files into a dedicated 'settings' directory and updating include paths
