+++ src/settings/recentfiles.cpp
